+++ src/components/seller/ProductCreate.tsx
@@ -5,37 +5,13 @@ import {
   TextField,
   MenuItem,
   Button,
-  Stack,
-  IconButton,
   InputAdornment,
   OutlinedInput,
   Box,
   Divider,
-  Card,
-  CardActionArea,
-  CardMedia,
-  Badge,
-  Typography,
-  Dialog,
-  DialogTitle,
   ToggleButtonGroup,
   ToggleButton,
 } from '@mui/material';
-
-import CloudUploadIcon from '@mui/icons-material/CloudUpload';
-import CloseIcon from '@mui/icons-material/Close';
-import CircularProgress from '@mui/material/CircularProgress';
-
-import { api } from '../../api/api';
-import { CATEGORY, QUALITY } from '../../constants/index';
-import { IProduct } from '../../type';
-
-import {
-  validateProductName,
-  validateProductContent,
-  validateProductPrice,
-  validateProductShippingFees,
-} from '../../lib/validation';
 import { FormLabel, ListItemDecorator, Menu, Textarea } from '@mui/joy';
 import {
   FormatBold,
@@ -43,8 +19,18 @@ import {
   KeyboardArrowDown,
 } from '@mui/icons-material';
 import Check from '@mui/icons-material/Check';
-import { useMutation } from 'react-query';
-const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
+
+import { api } from '../../api/api';
+import { CATEGORY, QUALITY } from '../../constants/index';
+import { IProduct } from '../../type';
+import {
+  validateProductName,
+  validateProductContent,
+  validateProductPrice,
+  validateProductShippingFees,
+} from '../../lib/validation';
+import FileUpload from '../FileUpload';
+
 const initCreateData = {
   price: 0,
   shippingFees: 0,
@@ -64,19 +50,20 @@ const initCreateData = {
     sort: 1,
   },
 };
+
+interface FilePreview {
+  id: string;
+  path: string;
+}
+
 export default function ProductCreate() {
   const userId = localStorage.getItem('_id');
 
   const [productData, setProductData] = useState<IProduct>(initCreateData);
   const [isValid, setIsValid] = useState(true);
-  const [filePreview, setFilePreview] = useState<
-    { id: string; path: string }[]
-  >([]);
   const [italic, setItalic] = useState(false);
   const [fontWeight, setFontWeight] = useState('normal');
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const [openDialog, setOpenDialog] = useState(false);
-  const [selectedImage, setSelectedImage] = useState(null);
   const [selectedCategory, setSelectedCategory] = useState('');
   const [selectedQuality, setSelectedQuality] = useState<string | number>('');
 
@@ -84,58 +71,17 @@ export default function ProductCreate() {
   const [priceError, setPriceError] = useState('');
   const [shippingFeesError, setShippingFeesError] = useState('');
   const [contentError, setContentError] = useState('');
+  const [productFiles, setProductFiles] = useState<FilePreview[]>([]);
 
-  // 사진 클릭시 팝업
-  const handleClickOpen = (image: any) => {
-    setSelectedImage(image);
-    setOpenDialog(true);
-  };
-
-  const handleClose = () => {
-    setOpenDialog(false);
+  const handleFilesChange = (files: FilePreview[]) => {
+    setProductFiles(files);
+    setProductData((prevData) => ({
+      ...prevData,
+      mainImages: [...prevData.mainImages, ...files],
+    }));
   };
 
   const navigate = useNavigate();
-
-  //가격, 배송료, 상품명, 상품 설명 상태값 업데이트
-  const handleAllChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-    const { name, value } = event.target;
-    if (name === 'price' || name === 'shippingFees') {
-      const numericValue = parseFloat(value);
-      setProductData((prev) => ({ ...prev, [name]: numericValue }));
-    } else {
-      setProductData((prev) => ({ ...prev, [name]: value }));
-    }
-  };
-
-  //상품 등록 유효성 검사
-  useEffect(() => {
-    if (!validateProductName(productData.name)) {
-      setIsValid(false);
-      setNameError('상품명은 2글자 이상 입력하세요.');
-    } else {
-      setIsValid(true);
-      setNameError('');
-    }
-    if (!validateProductPrice(productData.price)) {
-      setIsValid(false);
-      setPriceError('상품 가격은 정수로 입력하세요.');
-    } else {
-      setPriceError('');
-    }
-    if (!validateProductShippingFees(productData.shippingFees)) {
-      setIsValid(false);
-      setShippingFeesError('배송비는 정수로 입력하세요.');
-    } else {
-      setShippingFeesError('');
-    }
-    if (!validateProductContent(productData.content)) {
-      setIsValid(false);
-      setContentError('상품 설명을 10글자 이상 입력하세요.');
-    } else {
-      setContentError('');
-    }
-  }, [handleAllChange]);
 
   //뒤로가기
   const handleMoveBack = (e: React.MouseEvent<HTMLButtonElement>) => {
@@ -179,6 +125,46 @@ export default function ProductCreate() {
     }
   };
 
+  //가격, 배송료, 상품명, 상품 설명 상태값 업데이트
+  const handleAllChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+    const { name, value } = event.target;
+    if (name === 'price' || name === 'shippingFees') {
+      const numericValue = parseFloat(value);
+      setProductData((prev) => ({ ...prev, [name]: numericValue }));
+    } else {
+      setProductData((prev) => ({ ...prev, [name]: value }));
+    }
+  };
+
+  //상품 등록 유효성 검사
+  useEffect(() => {
+    if (!validateProductName(productData.name)) {
+      setIsValid(false);
+      setNameError('상품명은 2글자 이상 입력하세요.');
+    } else {
+      setIsValid(true);
+      setNameError('');
+    }
+    if (!validateProductPrice(productData.price)) {
+      setIsValid(false);
+      setPriceError('상품 가격은 정수로 입력하세요.');
+    } else {
+      setPriceError('');
+    }
+    if (!validateProductShippingFees(productData.shippingFees)) {
+      setIsValid(false);
+      setShippingFeesError('배송비는 정수로 입력하세요.');
+    } else {
+      setShippingFeesError('');
+    }
+    if (!validateProductContent(productData.content)) {
+      setIsValid(false);
+      setContentError('상품 설명을 10글자 이상 입력하세요.');
+    } else {
+      setContentError('');
+    }
+  }, [handleAllChange]);
+
   //상품데이터 form submit
   const productAllSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -196,63 +182,9 @@ export default function ProductCreate() {
     }
   };
 
-  // 파일 업로드
-  const uploadFileMutation = useMutation<any, Error, FormData>(
-    (newFiles) => {
-      return api.uploadFile(newFiles);
-    },
-    {
-      onSuccess: (response) => {
-        const files = response.data.files || [response.data.file];
-        const newPreviews = files.map((file: File) => ({
-          id: file.name,
-          path: `${API_BASE_URL}${(file as any).path}`,
-        }));
-        setFilePreview((currentPreviews) => [
-          ...currentPreviews,
-          ...newPreviews,
-        ]);
-        setProductData((currentData) => ({
-          ...currentData,
-          mainImages: [...currentData.mainImages, ...newPreviews],
-        }));
-      },
-      onError: (error) => {
-        console.error('Error uploading file:', error);
-      },
-    },
-  );
-
-  const handleFileUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault();
-    const files = e.target.files;
-    if (files) {
-      const formData = new FormData();
-      Array.from(files).forEach((file) => {
-        formData.append('attach', file as Blob);
-      });
-      uploadFileMutation.mutate(formData);
-    }
-  };
-
-  //파일 삭제
-  const handleFileRemove = (indexToRemove: string) => {
-    let updatedFilePreview = [...filePreview];
-    updatedFilePreview = updatedFilePreview.filter(
-      (item) => item.id !== indexToRemove,
-    );
-    setFilePreview(updatedFilePreview);
-
-    setProductData({
-      ...productData,
-      mainImages: updatedFilePreview,
-    });
-  };
-
-  console.log('data', productData);
-
   return (
     <form onSubmit={productAllSubmit}>
+      {/* 제목 */}
       <Box
         sx={{
           position: 'sticky',
@@ -269,6 +201,7 @@ export default function ProductCreate() {
         <Divider />
       </Box>
 
+      {/* 카테고리, 품질, 상품명, 가격, 배송비, 설명, 사진 업로드 */}
       <Box
         sx={{
           display: 'flex',
@@ -347,6 +280,7 @@ export default function ProductCreate() {
           ></OutlinedInput>
           {!isValid ? <div style={{ color: 'red' }}>{priceError}</div> : <></>}
         </Box>
+
         <Box>
           <FormLabel sx={{ fontSize: 'medium' }}>배송비:</FormLabel>
           <OutlinedInput
@@ -432,107 +366,7 @@ export default function ProductCreate() {
           )}
         </Box>
 
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: 'column',
-            gap: 2,
-            border: '1px solid #e0e0e0',
-            borderRadius: '4px',
-            padding: '16px',
-          }}
-        >
-          <Button
-            component="label"
-            variant="outlined"
-            startIcon={<CloudUploadIcon />}
-            disabled={filePreview.length >= 10}
-          >
-            사진 업로드
-            <input
-              hidden
-              accept="image/*"
-              multiple
-              type="file"
-              onChange={handleFileUpload}
-            />
-          </Button>
-          {uploadFileMutation.isLoading && (
-            <Box sx={{ display: 'flex', justifyContent: 'center' }}>
-              <CircularProgress />
-            </Box>
-          )}
-          <Stack direction="row" spacing={2}>
-            {filePreview.map((image, index) => (
-              <Badge
-                badgeContent={
-                  <CloseIcon
-                    onClick={() => handleFileRemove(image.id)}
-                    sx={{
-                      fontSize: '1rem',
-                      color: 'white',
-                      backgroundColor: 'red',
-                      borderRadius: '50%',
-                      cursor: 'pointer',
-                      '&:hover': {
-                        backgroundColor: 'darkred',
-                      },
-                    }}
-                  />
-                }
-                overlap="circular"
-                anchorOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                key={image.id}
-              >
-                {' '}
-                <div onClick={() => handleClickOpen(image.path)} key={image.id}>
-                  <Card sx={{ width: 90, height: 90 }}>
-                    <CardActionArea>
-                      <CardMedia
-                        component="img"
-                        height="90"
-                        image={image.path}
-                        alt={`Preview ${index + 1}`}
-                      />
-                    </CardActionArea>
-                  </Card>
-                </div>
-              </Badge>
-            ))}
-          </Stack>
-          <Box sx={{ alignSelf: 'center' }}>
-            {filePreview.length > 0 && (
-              <Typography variant="caption">{`${filePreview.length}/10`}</Typography>
-            )}
-            <Dialog onClose={handleClose} open={openDialog}>
-              <DialogTitle>
-                사진 미리보기
-                <IconButton
-                  aria-label="close"
-                  onClick={handleClose}
-                  sx={{
-                    position: 'absolute',
-                    right: 8,
-                    top: 8,
-                    color: (theme) => theme.palette.grey[500],
-                  }}
-                >
-                  <CloseIcon />
-                </IconButton>
-              </DialogTitle>
-              {selectedImage && (
-                <img
-                  src={selectedImage}
-                  alt={`Preview` + selectedImage}
-                  style={{ width: '100%' }}
-                />
-              )}
-            </Dialog>
-          </Box>
-        </Box>
+        <FileUpload onFilesChange={handleFilesChange} />
       </Box>
 
       <Box
